feat: return only the latest rule version from EprRuleFetcher
The package contains all versions of all rules, and no nice 'latest'
flag by which we could select only the latest versions of these rules.
So, we build a map of the rules by ID, keeping track of the latest
version of each, and then use that list to select the final list of
rules.