chore: fixing some types by making a class member public
There's no way to make a private member without using a class, so the
simplest solution here is just to make the member public.

I did have an alternate fix here where I made a thunk returning the actual client as an
object (closed over `fetch` instead of having it as an instance member).
But, as you might be able to tell from the length of explanation, that
was a lot more complex and I didn't like it as much.