chore: extracts retrieval of 'latest' rules into helper
Well, this was fun. I always end up playing with this kind of
abstraction way too much, and this the textbook definition of
overengineering. 86 lines of new code and 106 lines of tests to remove
14 LOC from another function :).

But: now it works, and it's tested, and gosh darnit I had fun doing it.
Thanks @elastic for giving us these ON weeks <3.